feat: 카드 입력 섹션의 상태 관리 및 유효성 검사 로직을 훅으로 분리
